+++ src/App.js
@@ -2,7 +2,7 @@ import './App.css';
 
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
 
-import { ArticlesProvider } from "./contexts/Articles";
+import { ArticlesProvider } from "./contexts/ContextArticles";
 
 import Home from "./components/home/Home"
 import Article from './components/articles/articles/Article';
